Move combine_polys and related structs into utils.rs
